Retry the whole transaction when there's an error
From reading the datasheet, it really doesn't look like issuing a second
read if there's an error on the first one is guaranteed to work. It
looks safer to reissue the whole thing by sending the commend and
reading 6-bytes again.

This means that using Circuits.I2C to retry writes or reads might not do
the right thing. It looks safer to retry the transaction as a whole.