Reject more known bad values
This adds more bad values that were seen on an SHT41. The 0x8002/0x8002
response would sometimes oscillate with 0x8003, but it was stuck on the
sensor. This led to constant 40C reports when the temperature wasn't
close to that.

This change just rejects any combinations where the temperature and
humidity are the magic values to head off other variations.